# api/openapi.yaml
openapi: 3.1.0
info:
  title: Aifory Stay Booking API
  version: 0.1.0
  description: Provider-neutral contract for hotel discovery and booking flows.
servers:
  - url: http://127.0.0.1:8080
    description: Local mock provider
tags:
  - name: system
  - name: locations
  - name: hotels
paths:
  /healthz:
    get:
      tags: [system]
      operationId: getHealth
      responses:
        "200":
          description: Service is ready
          content:
            application/json:
              schema:
                type: object
                required: [status, service, provider, time]
                properties:
                  status: { type: string, const: ok }
                  service: { type: string }
                  provider: { type: string }
                  time: { type: string, format: date-time }
  /v1/locations/suggest:
    get:
      tags: [locations]
      operationId: suggestLocations
      parameters:
        - in: query
          name: q
          schema: { type: string, maxLength: 120 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 20, default: 8 }
      responses:
        "200":
          description: Matching destinations
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Location" }
  /v1/hotels/search:
    post:
      tags: [hotels]
      operationId: searchHotels
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SearchRequest" }
      responses:
        "200":
          description: Hotel search results
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SearchResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /v1/hotels/{hotelID}:
    get:
      tags: [hotels]
      operationId: getHotel
      parameters:
        - in: path
          name: hotelID
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Hotel details
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Hotel" }
        "404":
          description: Hotel not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
components:
  responses:
    BadRequest:
      description: Invalid JSON request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    UnprocessableEntity:
      description: Search parameters are logically invalid
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    Money:
      type: object
      required: [amountMinor, currency]
      properties:
        amountMinor: { type: integer, format: int64, minimum: 0 }
        currency: { type: string, minLength: 3, maxLength: 3, example: RUB }
    Location:
      type: object
      required: [id, type, name, country, countryCode, latitude, longitude, propertyCount]
      properties:
        id: { type: string }
        type: { type: string, enum: [city, region, country, airport] }
        name: { type: string }
        country: { type: string }
        countryCode: { type: string }
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
        propertyCount: { type: integer }
    RoomOccupancy:
      type: object
      required: [adults, childrenAges]
      properties:
        adults: { type: integer, minimum: 1, maximum: 8 }
        childrenAges:
          type: array
          maxItems: 6
          items: { type: integer, minimum: 0, maximum: 17 }
    SearchFilters:
      type: object
      properties:
        stars:
          type: array
          uniqueItems: true
          items: { type: integer, minimum: 1, maximum: 5 }
        maxPriceMinor: { type: integer, format: int64, minimum: 0 }
        minReviewScore: { type: number, minimum: 0, maximum: 10 }
        amenities:
          type: array
          items: { type: string }
        refundableOnly: { type: boolean, default: false }
    SearchRequest:
      type: object
      additionalProperties: false
      required: [query, checkIn, checkOut, rooms, currency, sort, filters, limit]
      properties:
        query: { type: string, maxLength: 160 }
        checkIn: { type: string, format: date }
        checkOut: { type: string, format: date }
        rooms:
          type: array
          minItems: 1
          maxItems: 8
          items: { $ref: "#/components/schemas/RoomOccupancy" }
        currency: { type: string, example: RUB }
        sort:
          type: string
          enum: [recommended, price_asc, price_desc, rating_desc, distance_asc]
        filters: { $ref: "#/components/schemas/SearchFilters" }
        limit: { type: integer, minimum: 1, maximum: 100, default: 24 }
        cursor: { type: string }
    HotelSummary:
      type: object
      required: [id, name, slug, city, country, latitude, longitude, stars, reviewScore, reviewCount, distanceToCenterKm, imageUrl, amenities, badges, pricePerNight, totalPrice, availableRooms, refundable, mealPlan]
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        city: { type: string }
        country: { type: string }
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
        stars: { type: integer }
        reviewScore: { type: number }
        reviewCount: { type: integer }
        distanceToCenterKm: { type: number }
        imageUrl: { type: string, format: uri }
        amenities:
          type: array
          items: { type: string }
        badges:
          type: array
          items: { type: string }
        pricePerNight: { $ref: "#/components/schemas/Money" }
        totalPrice: { $ref: "#/components/schemas/Money" }
        availableRooms: { type: integer }
        refundable: { type: boolean }
        mealPlan: { type: string }
    Hotel:
      type: object
      required: [id, name, slug, description, city, country, countryCode, latitude, longitude, stars, reviewScore, reviewCount, distanceToCenterKm, imageUrl, images, amenities, badges, availableRooms, refundable, mealPlan]
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        city: { type: string }
        country: { type: string }
        countryCode: { type: string }
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
        stars: { type: integer }
        reviewScore: { type: number }
        reviewCount: { type: integer }
        distanceToCenterKm: { type: number }
        imageUrl: { type: string, format: uri }
        images:
          type: array
          items: { type: string, format: uri }
        amenities:
          type: array
          items: { type: string }
        badges:
          type: array
          items: { type: string }
        availableRooms: { type: integer }
        refundable: { type: boolean }
        mealPlan: { type: string }
    SearchFacets:
      type: object
      required: [minPriceMinor, maxPriceMinor, starCounts, amenityCounts]
      properties:
        minPriceMinor: { type: integer, format: int64 }
        maxPriceMinor: { type: integer, format: int64 }
        starCounts:
          type: object
          additionalProperties: { type: integer }
        amenityCounts:
          type: object
          additionalProperties: { type: integer }
    SearchResponse:
      type: object
      required: [searchId, items, total, facets, expiresAt]
      properties:
        searchId: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/HotelSummary" }
        total: { type: integer }
        nextCursor: { type: string }
        facets: { $ref: "#/components/schemas/SearchFacets" }
        expiresAt: { type: string, format: date-time }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }